% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_datasets.R
\name{available_datasets}
\alias{available_datasets}
\title{Return a list of the available datasets}
\usage{
available_datasets()
}
\value{
List of available datasets
}
\description{
Return a list of the available datasets
}
\examples{
\dontrun{
available_datasets()
}
}
