% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_datasets.R
\name{fetch_datasets}
\alias{fetch_datasets}
\title{Return the local locations of datasets listed in listdatasets (downloading them from the internet if needed)}
\usage{
fetch_datasets(
  listofdatasets,
  locations =
    "https://github.com/chrplr/openlexicon/tree/master/datasets-info/locations.toml"
)
}
\arguments{
\item{listofdatasets}{A list of datasets to be loaded.}

\item{locations}{A list of locations (hosted online).}
}
\value{
Local locations of datasets listed in listdatasets.
}
\description{
Return the local locations of datasets listed in listdatasets (downloading them from the internet if needed)
}
\examples{
\dontrun{
fetch_datasets(c('Lexique3', 'Voisins', 'Anagrammes') )
}
}
