% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_datasets.R
\name{get_all_datasets}
\alias{get_all_datasets}
\title{Download all the datasets (only if not already downloaded)}
\usage{
get_all_datasets()
}
\value{
Local locations of all datasets listed in locations.
}
\description{
Download all the datasets (only if not already downloaded)
}
\examples{
\dontrun{
get_all_datasets()
}
}
