% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.home.R
\name{get_data.home}
\alias{get_data.home}
\title{Returning the path of the local folder where to put datasets}
\usage{
get_data.home()
}
\value{
A path for the datasets folder.
}
\description{
Returning the path of the local folder where to put datasets
}
\examples{
get_data.home()
}
