% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_from_json.R
\name{get_dataset_from_json}
\alias{get_dataset_from_json}
\title{Downloading (only if needed) the dataset 'filename' indicated from the json description file}
\usage{
get_dataset_from_json(json_url, local_rds_name)
}
\arguments{
\item{json_url}{URL to json file.}

\item{local_rds_name}{Name of the local rds file.}
}
\value{
The path to the local version of the dataset.
}
\description{
Downloading (only if needed) the dataset 'filename' indicated from the json description file
}
\examples{
\dontrun{
get_dataset_from_json(
    json_url = "http://www.lexique.org/databases/_json/Lexique383.json",
    local_rds_name = "Lexique383"
    )
}
}
