% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_from_json.R
\name{get_info_from_json}
\alias{get_info_from_json}
\title{Extract information fields from a json file}
\usage{
get_info_from_json(json_url)
}
\arguments{
\item{json_url}{URL to json file.}
}
\value{
A list of information.
}
\description{
Extract information fields from a json file
}
\examples{
get_info_from_json(json_url = "http://www.lexique.org/databases/_json/anagrammes.json")
}
