% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_log.home.R
\name{get_log.home}
\alias{get_log.home}
\title{Return the path of the folder where to put logs}
\usage{
get_log.home(app_name)
}
\arguments{
\item{app_name}{A character vector with, at most, one element.}
}
\value{
A path to the logs folder.
}
\description{
Return the path of the folder where to put logs
}
\examples{
get_log.home
}
