% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsv_to_rds.R
\name{tsv_to_rds}
\alias{tsv_to_rds}
\title{Load in memory the rds files listed in list_rds, in variables with the names matching names(list_rds).}
\usage{
tsv_to_rds(tsv_url, table_name)
}
\arguments{
\item{tsv_url}{URL to the original tsv file.}

\item{table_name}{Name of the output rds file (without extension).}
}
\value{
Returns a rds file.
}
\description{
Load in memory the rds files listed in list_rds, in variables with the names matching names(list_rds).
}
\examples{
\dontrun{
tsv_to_rds(tsv_url)
}
}
