% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_datasets.R
\name{fetch_datasets}
\alias{fetch_datasets}
\title{Returning a list of datasets (downloading those datasets from the internet if needed)}
\usage{
fetch_datasets(datasets = c("Lexique3", "Voisins", "Anagrammes"))
}
\arguments{
\item{datasets}{A list of datasets to be loaded.}
}
\value{
Local locations of datasets.
}
\description{
Returning a list of datasets (downloading those datasets from the internet if needed)
}
\examples{
\dontrun{
fetch_datasets(c("Lexique3", "Voisins", "Anagrammes") )
}
}
