% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_datasets.R
\name{get_all_datasets}
\alias{get_all_datasets}
\title{Download all the datasets listed in locations (only if not already downloaded)}
\usage{
get_all_datasets(
  locations =
    "https://github.com/chrplr/openlexicon/tree/master/datasets-info/locations.toml"
)
}
\arguments{
\item{locations}{A list of locations.}
}
\value{
Local locations of all datasets listed in locations.
}
\description{
Download all the datasets listed in locations (only if not already downloaded)
}
\examples{
\dontrun{
loc <- "https://github.com/chrplr/openlexicon/tree/master/datasets-info/locations.toml"
get_all_datasets(loc)
}
}
